Move `FeedImageDataStoreSpy` to share scope
